fix: RestTemplateConfig가 2개 정의되어 있는 현상
