--- src/main/java/net/app/savable/global/config/redis/RestTemplateConfig.java
@@ -1,28 +1,0 @@
-package net.app.savable.global.config.redis;
-
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.BufferingClientHttpRequestFactory;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
-import org.springframework.http.converter.StringHttpMessageConverter;
-import org.springframework.web.client.RestTemplate;
-
-import java.nio.charset.StandardCharsets;
-
-@Configuration
-public class RestTemplateConfig {
-
-        @Bean
-        public RestTemplate restTemplate(RestTemplateBuilder restTemplateBuilder) {
-            return restTemplateBuilder
-                    .requestFactory(
-                            () -> new BufferingClientHttpRequestFactory(
-                                    new SimpleClientHttpRequestFactory()
-                            )
-                    ).additionalMessageConverters(
-                            new StringHttpMessageConverter(
-                                    StandardCharsets.UTF_8
-                            )).build();
-        }
-}
